localization work

+++ app/src/main/java/com/teamx/hatlyDriver/data/dataclasses/SignUp/RegisterDara.kt
@@ -8,5 +8,5 @@ data class RegisterData(
     val name: String,
     val role: String,
     val verified: Boolean,
-    val wallet: Int
+    val wallet: Double
 )
